+++ src/IMU_LSM6DS3TR_C.cpp
@@ -325,7 +325,7 @@ int IMU_LSM6DS3TR_C::init(uint32_t targetOutputDataRateHz, gyro_sensitivity_e gy
     delayMs(1);
 
     // return the gyro sample rate actually set
-    return _gyroSampleRateHz;
+    return static_cast<int>(_gyroSampleRateHz);
 }
 
 void IMU_LSM6DS3TR_C::setInterruptDriven()
